Created a separate area for extensions.
Extensions are feature not (yet?) available in the official mdl project. They are not to be mixed with the official set of components to keep things clear of what to be expected from each set of components. Since the components and extensions will live inside two different pages, it was necessary to move the related demo code into separated modules. The demo application, now runs two distinct modules.