Appying template over the button demo html page
Making use of the freshly built front end workflow to implement templates over the demo pages. The source code of the demo is wired to the page avoiding the maintenance nightmare that would result from the required updates on the demo for each update on the source code of the demo.