GWT Tests are now performed through a dedicated GWT module
The test module inherits from the tested library module and provide the context needed to fully test the library. It is responsible for importing the required mdl css and javascript. It includes permutations cut down and invokes the appropriate linker (unfortunatly deprecated) to be able to inject javascipt into the test page.