Removed tab-size css property in code snippets
The project is following the checkstyle standard, which means there is not a single tab character in the whole java codebase of the project anymore. Therefore, code snippets do not need to override the ugly 8 spaces defaut tab width to save horizontal space in the demos. This is a convenient situation because tab-size is not implemented by all browsers. Which makes it one less non-standard css property to worry about.