Removed the extra logging jibberish when testing the library
The inclusion of material.css in the testing host page accounts for most of the logging activity. It is not needed to show all the warning emmitted by htmlunit and at the same time there is no solution to shut it down. The only option left is to drop the css altogether. It is possible to reactivate it when needed. The other logging source is the library itself. Here again, it is better to disable logging when everything is fine.