Relaxing the aggressive "design for extension" checkstyle rule
This rule is really impractical to commit to as the benefit of overriding methods in a sub-type outweights the danger of that practice if done carefully. In many scenarios, it is just unnatural to rely on explicitly placed hooks to provide a way to change the original behavior and sometimes makes the design more complicated to understand and maintain. The hook workaround may work in many situations but fails to provide a clean solution when it comes to completly change the methods behavior in the subtype.