Added a tooltip component impementation
This implementation is a little different from the MDL one. The MDL version uses z-indexes to arbitrarely set the depth of the tooltip while this implementation uses popup panels to add the tooltip as the last child of the body tag exactly when the tooltip is shown to the user to insure it is always displayed above everything. Unlike the original implementation which relies on an arbitrarely fixed value of z-index to set the depth of the tooltip, this implementation is cleaner and more predictable as it prevents other compoments to overlap the tooltip caused by z-index outbidding.