Fixed irrelevant code on the radio button
The original implementation of the sinkEvents method on the widget type set the sinking on the container element of the widget. Then the GWT checkbox came in. The GWT checkbox inherits the widget and sets the sinking on the input rather than the container. Then the mdl implementation was introduced that followed the same design pattern as the GWT checkbox. Then the GWT RadioButton was introduced which extended the mdl checkbox. The problem is that when events are sunk on the input (behavior inherited from the mdl checkbox) change handlers where not captured if the radio label is clicked. So the mdl radio had to override the sinkEvents to sink on the container in the same way as the widget. Later (in the last commit) it was discovered that sinking on the input does not suit the mdl checkbox either. So in conclusion, it turns out that this whole GWT idea of sinking on the input is not suitable for the mdl version that do not display the input element at all.